@@ -12,26 +12,26 @@
 		"format:prettier": "prettier --write ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-static": "^1.0.0-next.4",
-		"@sveltejs/kit": "next",
-		"@typescript-eslint/eslint-plugin": "^4.19.0",
-		"@typescript-eslint/parser": "^4.19.0",
-		"eslint": "^7.22.0",
+		"@sveltejs/adapter-static": "^1.0.0-next.13",
+		"@sveltejs/kit": "^1.0.0-next.113",
+		"@typescript-eslint/eslint-plugin": "^4.26.0",
+		"@typescript-eslint/parser": "^4.26.0",
+		"eslint": "^7.27.0",
 		"eslint-config-prettier": "^8.1.0",
 		"eslint-plugin-svelte3": "^3.2.0",
 		"markdown-it": "^12.0.4",
 		"markdown-it-anchor": "^7.1.0",
 		"markdown-it-table-of-contents": "^0.5.2",
 		"normalize.css": "^8.0.1",
-		"prettier": "~2.2.1",
-		"prettier-plugin-svelte": "^2.2.0",
-		"slugify": "^1.5.0",
-		"svelte": "^3.29.0",
-		"svelte-check": "^1.4.0",
-		"svelte-preprocess": "^4.7.1",
+		"prettier": "^2.3.0",
+		"prettier-plugin-svelte": "^2.3.0",
+		"slugify": "^1.5.3",
+		"svelte": "^3.38.2",
+		"svelte-check": "^2.0.0",
+		"svelte-preprocess": "^4.7.3",
 		"tslib": "^2.0.0",
 		"typescript": "^4.0.0",
-		"vite": "^2.2.3",
+		"vite": "^2.3.6",
 		"vite-imagetools": "^3.4.0",
 		"vite-plugin-markdown": "^2.0.2"
 	},
